+++ src/ControllerMethodArgs_test.ts
@@ -126,6 +126,53 @@ Deno.test("ControllerMethodArgs Decorator - Standard strategy - declaring body, 
   assertEquals(enhancedHandlerRetVal, { "i.am": "deep" });
 });
 
+Deno.test("ControllerMethodArgs Decorator - Standard strategy - declaring body, param, query, headers", async () => {
+  let enhancedHandler;
+  let enhancedHandlerRetVal;
+  try {
+    enhancedHandler = ControllerMethodArgs("body", "param", "query", "headers")(
+      function testHandler(
+        _body: unknown,
+        _param: unknown,
+        _query: unknown,
+        headers: unknown,
+      ) {
+        // @TODO consider mocking `body`, `param`, and `query` and assert for
+        // their parsed values
+        return {
+          "i.am.also": "deep",
+          "my.headers.include": headers,
+        };
+      },
+      {
+        name: "testHandler",
+      } as ClassMethodDecoratorContext,
+    );
+  } catch (e) {
+    throw new Error(`test case should not have thrown ${e.message}`);
+  }
+  assertEquals(typeof enhancedHandler, "function");
+
+  try {
+    const ctx = createMockContext();
+    Object.defineProperty(ctx.request, "body", {
+      get: () => createMockRequestBody("json"),
+    });
+    Object.defineProperty(ctx.request, "headers", {
+      value: new Map([["x-foo", "bar"]]),
+    });
+
+    enhancedHandlerRetVal = await enhancedHandler(ctx);
+  } catch (e) {
+    throw new Error(`test case should not have thrown ${e.message}`);
+  }
+
+  assertEquals(enhancedHandlerRetVal, {
+    "i.am.also": "deep",
+    "my.headers.include": { "x-foo": "bar" },
+  });
+});
+
 Deno.test("ControllerMethodArgs Decorator - Standard strategy - declaring param, body", async () => {
   let enhancedHandler;
   let enhancedHandlerRetVal;
@@ -306,6 +353,59 @@ Deno.test("ControllerMethodArgs Decorator - CloudflareWorker strategy - declarin
   }
 
   assertEquals(enhancedHandlerRetVal, { "i.am": "deep.too" });
+});
+
+Deno.test("ControllerMethodArgs Decorator - CloudflareWorker strategy - declaring body, param, query, headers", async () => {
+  const methodDescriptor = {
+    value: function testHandler(
+      _body: unknown,
+      _param: unknown,
+      _query: unknown,
+      headers: unknown,
+    ) {
+      return {
+        "i.am.also": "deep.ya",
+        "my.headers.include": headers
+      };
+    },
+    writable: true,
+    enumerable: false,
+    configurable: true,
+  };
+  let decoratorRetVal;
+  try {
+    decoratorRetVal = ControllerMethodArgs("body", "param", "query", "headers")(
+      {},
+      "testHandler",
+      methodDescriptor,
+    );
+  } catch (e) {
+    throw new Error(`test case should not have thrown ${e.message}`);
+  }
+  assertEquals(decoratorRetVal, undefined);
+
+  // by this point, `methodDescriptor.value` has been written with a new value
+  // which is the enhanced `testHandler`
+  let enhancedHandlerRetVal;
+  try {
+    // deno-lint-ignore ban-types
+    const enhancedHandler: Function = methodDescriptor.value;
+    const ctx = createMockContext();
+    Object.defineProperty(ctx.request, "body", {
+      get: () => createMockRequestBody("json"),
+    });
+    Object.defineProperty(ctx.request, "headers", {
+      value: new Map([["x-foo", "bar"]])
+    })
+    enhancedHandlerRetVal = await enhancedHandler(ctx);
+  } catch (e) {
+    throw new Error(`test case should not have thrown ${e.message}`);
+  }
+
+  assertEquals(enhancedHandlerRetVal, {
+    "i.am.also": "deep.ya",
+    "my.headers.include": { "x-foo": "bar" },
+  });
 });
 
 Deno.test("ControllerMethodArgs Decorator - CloudflareWorker strategy - declaring param, body", async () => {